update fuc: readFiles

+++ get_cluster.h
@@ -25,7 +25,7 @@ typedef enum ERROR_TABLE_FAT
     EOF_CLUSTER = 0xfff
 } ERROR_FAT;
 
-uint32_t getNextCluster(uint32_t startCluster, FILE *file);
+uint16_t getNextCluster(uint16_t startCluster, FILE *file);
 uint32_t getAddressCluster(const BootSector_t *bs, uint32_t startCluster);
 
 error_code_t getEntry(FILE *fp, const BootSector_t *bs, DirectoryEntry_t *entryOut);
